# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MIRA Foundation",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/memory-management/status": {
      "get": {
        "tags": [
          "local memory management"
        ],
        "summary": "Memory Management Status",
        "operationId": "memory_management_status_api_v1_memory_management_status_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemoryManagementStatusView"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/memory-management/entries": {
      "get": {
        "tags": [
          "local memory management"
        ],
        "summary": "List Memory Entries",
        "operationId": "list_memory_entries_api_v1_memory_management_entries_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 20,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 64
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemoryManagementPageView"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/memory-management/operations": {
      "post": {
        "tags": [
          "local memory management"
        ],
        "summary": "Apply Memory Operation",
        "operationId": "apply_memory_operation_api_v1_memory_management_operations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MemoryManagementOperationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemoryManagementOperationView"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "message",
          "request_id"
        ],
        "title": "ErrorResponse"
      },
...
      "MemoryManagementEntryView": {
        "properties": {
          "entry_id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[^\\s\\x00-\\x1f\\x7f]+$",
            "title": "Entry Id"
          },
          "text": {
            "type": "string",
            "maxLength": 16384,
            "minLength": 1,
            "title": "Text"
          },
          "kind": {
            "type": "string",
            "enum": [
              "episodic",
              "boundary"
            ],
            "title": "Kind"
          },
          "source": {
            "type": "string",
            "const": "user_statement",
            "title": "Source",
            "default": "user_statement"
          },
          "source_version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Source Version"
          },
          "recorded_at": {
            "type": "string",
            "maxLength": 64,
            "minLength": 20,
            "title": "Recorded At"
          },
          "active": {
            "type": "boolean",
            "title": "Active"
          },
          "forget_event_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128,
                "minLength": 1,
                "pattern": "^[^\\s\\x00-\\x1f\\x7f]+$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Forget Event Id"
          },
          "forgotten_at": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64,
                "minLength": 20
              },
              {
                "type": "null"
              }
            ],
            "title": "Forgotten At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "entry_id",
          "text",
          "kind",
          "source_version",
          "recorded_at",
          "active",
          "forget_event_id",
          "forgotten_at"
        ],
        "title": "MemoryManagementEntryView"
      },
      "MemoryManagementOperationRequest": {
        "properties": {
          "operation_id": {
            "type": "string",
            "format": "uuid",
            "title": "Operation Id"
          },
          "expected_revision": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Expected Revision"
          },
          "operation": {
            "type": "string",
            "enum": [
              "record",
              "correct",
              "forget",
              "restore"
            ],
            "title": "Operation"
          },
          "confirmed": {
            "type": "boolean",
            "const": true,
            "title": "Confirmed"
          },
          "text": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 4096,
                "minLength": 1,
                "pattern": "\\S"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text"
          },
          "kind": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "episodic",
                  "boundary"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Kind"
          },
          "entry_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128,
                "minLength": 1,
                "pattern": "^[^\\s\\x00-\\x1f\\x7f]+$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Entry Id"
          },
          "forget_event_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128,
                "minLength": 1,
                "pattern": "^[^\\s\\x00-\\x1f\\x7f]+$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Forget Event Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "operation_id",
          "expected_revision",
          "operation",
          "confirmed"
        ],
        "title": "MemoryManagementOperationRequest"
      },
      "MemoryManagementOperationView": {
        "properties": {
          "status": {
            "type": "string",
            "const": "committed",
            "title": "Status",
            "default": "committed"
          },
          "operation_id": {
            "type": "string",
            "title": "Operation Id"
          },
          "revision": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Revision"
          },
          "entry_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128,
                "minLength": 1,
                "pattern": "^[^\\s\\x00-\\x1f\\x7f]+$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Entry Id"
          },
          "event_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128,
                "minLength": 1,
                "pattern": "^[^\\s\\x00-\\x1f\\x7f]+$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Event Id"
          },
          "replayed": {
            "type": "boolean",
            "title": "Replayed",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "operation_id",
          "revision"
        ],
        "title": "MemoryManagementOperationView"
      },
      "MemoryManagementPageView": {
        "properties": {
          "revision": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Revision"
          },
          "entries": {
            "items": {
              "$ref": "#/components/schemas/MemoryManagementEntryView"
            },
            "type": "array",
            "maxItems": 20,
            "title": "Entries"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "revision",
          "entries",
          "next_cursor"
        ],
        "title": "MemoryManagementPageView"
      },
      "MemoryManagementStatusView": {
        "properties": {
          "enabled": {
            "type": "boolean",
            "title": "Enabled"
          },
          "revision": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Revision"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "enabled",
          "revision"
        ],
        "title": "MemoryManagementStatusView"
      },
...
    }
  }
}